requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.0.6
django-bootstrap4==24.3
djangorestframework==3.15.2
filelock==3.15.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
setuptools==70.0.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.2
Werkzeug==3.0.3
wheel==0.43.0
